+++ src/main/java/at/porscheinformatik/seleniumcomponents/SeleniumComponent.java
@@ -2,6 +2,7 @@ package at.porscheinformatik.seleniumcomponents;
 
 import java.util.Objects;
 
+import org.openqa.selenium.NoSuchElementException;
 import org.openqa.selenium.WebElement;
 
 /**
@@ -38,12 +39,34 @@ public interface SeleniumComponent extends WebElementContainer
     }
 
     /**
-     * Returns true if a {@link WebElement} described by this component is ready and available (it must no be visible,
+     * Returns true if a {@link WebElement} described by this component is ready and available (it must not be visible,
      * though). This method has no timeout, it does not wait for the component to become existent.
      *
      * @return true if the component exists
      */
     boolean isReady();
+
+    /**
+     * Returns true if a {@link WebElement} described by this component is visible. By default it checks, if the element
+     * is visible. This method has no timeout, it does not wait for the component to become existent.
+     *
+     * @return true if the component is visible
+     */
+    default boolean isVisible()
+    {
+        try
+        {
+            return SeleniumUtils.retryOnStale(() -> {
+                WebElement element = element();
+
+                return element.isDisplayed();
+            });
+        }
+        catch (NoSuchElementException e)
+        {
+            return false;
+        }
+    }
 
     @Override
     String toString();